// src/plugins/query-runner/gql/generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Upload: { input: any; output: any; }
};
// ...
export type QueryResult = {
  __typename?: 'QueryResult';
  columns: Array<Scalars['String']['output']>;
  error?: Maybe<Scalars['String']['output']>;
  executionTime: Scalars['Int']['output'];
  rowCount: Scalars['Int']['output'];
  rows: Array<Scalars['JSON']['output']>;
};